# SSE: optimize WideToMany