<!--

Thank you for sending a pull request! Here are some tips:

1. If this is your first time, please read our contribution guide at https://github.com/grafana/grafana/blob/main/CONTRIBUTING.md

2. Ensure you include and run the appropriate tests as part of your Pull Request.

3. In a new feature or configuration option, an update to the documentation is necessary. Everything related to the documentation is under the docs folder in the root of the repository.

4. If the Pull Request is a work in progress, make use of GitHub's "Draft PR" feature and mark it as such.

5. If you can not merge your Pull Request due to a merge conflict, Rebase it. This gets it in sync with the main branch.

6. Name your PR as "<FeatureArea>: Describe your change", e.g. Alerting: Prevent race condition. If it's a fix or feature relevant for the changelog describe the user impact in the title. The PR title is used to auto-generate the changelog for issues marked with the "add to changelog" label.

7. If your PR content should be added to the What's New document for the next major or minor release, add the **add to what's new** label to your PR. Note that you should add this label to the main PR that introduces the feature; do not add this label to smaller PRs for the feature.

-->

**What is this feature?**
This PR refactors function `expr.WideToMany` that converts a frame to a collection of mathexp.Series (in general case).
It optimizes the function to not create a frame for each dimension and then call mathexp.SeriesFromFrame that mutates the frame if it does not have the correct schema (e.g. *float64 value field).
The PR replaces the function mathexp.SeriesFromFrame to a new function `mathexp.SeriesFromFrameFields` that accepts frame and indices of time and value fields. The function becomes idempotent.
Performance improvement was due to the fact that the function SeriesFromFrameFields does not create a copy of all fields in the series. Instead, it creates a new frame that has fields that point to the original frame, with one exception: if fields need to be converted to Time ( from NullableTime) and\or to NullableFloat64 (from any other). In this case, a new field is created and all values, config and labels are copied from the original field

**Why do we need this feature?**

It simplifies implementation of WideToMany and increases its performance 

before changes:
```
goos: windows
goarch: amd64
pkg: github.com/grafana/grafana/pkg/expr
cpu: 11th Gen Intel(R) Core(TM) i9-11900H @ 2.50GHz
BenchmarkWideToMany
BenchmarkWideToMany/time_and_*float64
BenchmarkWideToMany/time_and_*float64-16                   15796             76782 ns/op           73936 B/op       2117 allocs/op
BenchmarkWideToMany/*time_and_int64
BenchmarkWideToMany/*time_and_int64-16                      9488            130480 ns/op          136417 B/op       4197 allocs/op
```

after changes:
```
goos: windows
goarch: amd64
pkg: github.com/grafana/grafana/pkg/expr
cpu: 11th Gen Intel(R) Core(TM) i9-11900H @ 2.50GHz
BenchmarkWideToMany
BenchmarkWideToMany/time_and_*float64
BenchmarkWideToMany/time_and_*float64-16                   37693             32045 ns/op           25456 B/op        107 allocs/op
BenchmarkWideToMany/*time_and_int64
BenchmarkWideToMany/*time_and_int64-16                     19030             63392 ns/op           96176 B/op       2157 allocs/op
PASS
```

**Who is this feature for?**

Developers and users with big queries.

**Which issue(s) does this PR fix?**:

<!--

- Automatically closes linked issue when the Pull Request is merged.

Usage: "Fixes #<issue number>", or "Fixes (paste link of issue)"

-->

Fixes #

**Special notes for your reviewer:**

Please check that:
- [ ] It works as expected from a user's perspective.
- [ ] If this is a pre-GA feature, it is behind a feature toggle.
- [ ] The docs are updated, and if this is a [notable improvement](https://grafana.com/docs/writers-toolkit/writing-guide/contribute-release-notes/#how-to-determine-if-content-belongs-in-a-whats-new-document), it's added to our [What's New](https://grafana.com/docs/writers-toolkit/writing-guide/contribute-release-notes/) doc.