@leeoniya 2023-07-14
> Instead, it creates a new frame that has fields that point to the original frame, with one exception: if

if...